backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
redis==5.0.1

# Celery for task queue
celery==5.3.4
celery[redis]==5.3.4

# Scraping and parsing
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
feedparser==6.0.10
arxiv==2.1.0

# NLP and ML
sentence-transformers==2.2.2
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
datasets==2.16.1

# OpenAI for LLM agents
openai==1.6.1

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.5.3

# Monitoring
sentry-sdk[flask]==1.39.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

